rename: move authToken & provider package
