+++ src/main/java/com/example/backend/controller/LoginController.java
@@ -9,8 +9,8 @@ import com.example.backend.global.exception.user.UserInvalidInputException;
 import com.example.backend.global.exception.user.UserInvalidInputExceptionType;
 import com.example.backend.global.exception.UnAuthorizedException;
 import com.example.backend.global.exception.UnAuthorizedExceptionType;
-import com.example.backend.global.security.AuthToken;
-import com.example.backend.global.security.AuthTokenProvider;
+import com.example.backend.global.security.jwt.authToken.AuthToken;
+import com.example.backend.global.security.jwt.authToken.AuthTokenProvider;
 import com.example.backend.global.security.CookieProvider;
 import com.example.backend.global.utils.ResponseUtils;
 import com.example.backend.service.login.LoginService;
