+++ src/main/java/com/example/backend/controller/UserController.java
@@ -11,9 +11,9 @@ import com.example.backend.global.exception.ForbiddenException;
 import com.example.backend.global.exception.ForbiddenExceptionType;
 import com.example.backend.global.exception.user.UserInvalidInputException;
 import com.example.backend.global.exception.user.UserInvalidInputExceptionType;
-import com.example.backend.global.security.AuthToken;
+import com.example.backend.global.security.jwt.authToken.AuthToken;
 import com.example.backend.global.security.CustomUserDetails;
-import com.example.backend.global.security.AuthTokenProvider;
+import com.example.backend.global.security.jwt.authToken.AuthTokenProvider;
 import com.example.backend.global.security.CookieProvider;
 import com.example.backend.global.utils.ResponseUtils;
 import com.example.backend.service.user.SnsAPIService;
