+++ src/main/java/com/example/backend/global/security/jwt/authToken/AuthTokenProvider.java
@@ -1,8 +1,9 @@
-package com.example.backend.global.security;
+package com.example.backend.global.security.jwt.authToken;
 
 import com.example.backend.domain.User;
 import com.example.backend.global.exception.UnAuthorizedException;
 import com.example.backend.global.exception.UnAuthorizedExceptionType;
+import com.example.backend.global.security.CustomUserDetails;
 import com.example.backend.global.security.jwt.JwtConfig;
 import com.example.backend.global.security.jwt.JwtType;
 import com.fasterxml.jackson.core.JsonProcessingException;
