+++ src/main/java/com/example/backend/service/login/LoginServiceImpl.java
@@ -6,8 +6,8 @@ import com.example.backend.dto.login.DefaultLoginRequestDTO;
 import com.example.backend.global.exception.*;
 import com.example.backend.global.exception.user.UserInvalidInputException;
 import com.example.backend.global.exception.user.UserInvalidInputExceptionType;
-import com.example.backend.global.security.AuthToken;
-import com.example.backend.global.security.AuthTokenProvider;
+import com.example.backend.global.security.jwt.authToken.AuthToken;
+import com.example.backend.global.security.jwt.authToken.AuthTokenProvider;
 import com.example.backend.repository.UserRepository;
 import com.example.backend.service.user.SnsAPIService;
 import lombok.RequiredArgsConstructor;
